% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageviewer.R
\name{imageviewer}
\alias{imageviewer}
\title{imageviewer}
\usage{
imageviewer(data, width = NULL, height = NULL, elementId = NULL,
  options = list())
}
\arguments{
\item{data}{A matrix}

\item{width, height}{matrix dimensions}

\item{elementId}{HTML element Id}

\item{options}{list of other options (passed through to JS code)}
}
\value{
Plot matrix in html widget
}
\description{
Simple htmlwidgets matrix viewer with WebGL brightness/contrast
}
\examples{
# Create matrix
m <- matrix(rnorm(512 * 512, mean = 100, sd = 10), 512, 512)

# Plot
imageviewer(m)
}
